first final render

+++ Controllers/UtilisateurController.php
@@ -29,25 +29,29 @@ class UtilisateurController extends Controller
     public function createUserUt()
     {
         if ($_SERVER['REQUEST_METHOD'] === 'POST') {
-            $nom = htmlspecialchars(strip_tags(trim($_POST['name'])));
-            $prenom = htmlspecialchars(strip_tags(trim($_POST['surname'])));
+            $nom = htmlspecialchars($_POST['name']);
+            $prenom = htmlspecialchars($_POST['surname']);
             $mot_de_passe = trim($_POST['password']);
             $numero_telephone = trim($_POST['tel']);
-            $email = trim($_POST['email']);
-            $ville = htmlspecialchars(strip_tags(trim($_POST['city'])));
+            $email = filter_var(trim($_POST['email']), FILTER_SANITIZE_EMAIL);
+            $ville = htmlspecialchars(($_POST['city']));
             $role = 0;
 
-            $message[] = "";
+            if ($role != 0) {
+                $role = 0;
+            }
+
 
             if (!filter_var($email, FILTER_VALIDATE_EMAIL)) {
                 $message[] = "Email Invalide";
             }
-            if (strlen($mot_de_passe) < 8) {
-                $message[] = " Le mot de passe doit contenir au moins 8 caractères.";
-            }
 
             if (!preg_match('/^\+?[0-9]{10,15}$/', $numero_telephone)) {
                 $message[] = "Numéro de téléphone invalide";
+            }
+
+            if (strlen($mot_de_passe) < 8) {
+                $message[] = " Le mot de passe doit contenir au moins 8 caractères.";
             }
 
             if (!preg_match('/[A-Z]/', $mot_de_passe) || !preg_match('/[a-z]/', $mot_de_passe)) {
@@ -62,6 +66,8 @@ class UtilisateurController extends Controller
             if (empty($message)) {
                 $message[] = "";
             }
+
+            
             $mot_de_passe_hash = password_hash($mot_de_passe, PASSWORD_DEFAULT);
 
             $utilisateurModel = new UtilisateurModel();
@@ -83,11 +89,7 @@ class UtilisateurController extends Controller
 
             $utilisateurModel = new UtilisateurModel();
             if ($utilisateurModel->create($utilisateur)) {
-                $mailModel = new MailsModel();
-                $subject = "Bienvenue sur EliteDrive !";
-                $body = file_get_contents(__DIR__ . '/../Views/mails/subUt.php');
-                $to = $email;
-                $mailModel->sendMail($to, $subject, $body);
+                session_regenerate_id(true);
                 $_SESSION['id_utilisateur'] = $utilisateur->getId_utilisateur();
                 $_SESSION['email'] = $utilisateur->getEmail();
                 $_SESSION['numero_telephone'] = $utilisateur->getNumero_telephone();
@@ -95,8 +97,16 @@ class UtilisateurController extends Controller
                 $_SESSION['nom'] = $utilisateur->getNom();
                 $_SESSION['prenom'] = $utilisateur->getPrenom();
                 $_SESSION['ville'] = $utilisateur->getVille();
+                $_SESSION['crsf_token'] = bin2hex(random_bytes(32));
 
-                header('Location: index.php?controller=Home&action=showProfile');
+                $mailModel = new MailsModel();
+                $subject = "Bienvenue sur EliteDrive !";
+                $body = file_get_contents(__DIR__ . '/../Views/mails/subUt.php');
+                $to = $email;
+                $mailModel->sendMail($to, $subject, $body);
+                
+               
+                header('Location: index.php?controller=Utilisateur&action=showProfile');
                 exit();
             } else {
                 $message = "Erreur lors de l'inscription.";
@@ -123,19 +133,19 @@ class UtilisateurController extends Controller
     public function connectUser()
     {
         if ($_SERVER['REQUEST_METHOD'] === 'POST') {
-            $email = $_POST['email'];
-            $mot_de_passe = $_POST['password'];
-
+            $email = filter_input(INPUT_POST, 'email', FILTER_SANITIZE_EMAIL);
+            $mot_de_passe = trim($_POST['password']);
             $utilisateurModel = new UtilisateurModel();
-            $utilisateur = $utilisateurModel->connect($email, $mot_de_passe); 
-
+            $utilisateur = $utilisateurModel->connect($email, $mot_de_passe);
             if ($utilisateur) {
+                session_regenerate_id(true);
                 $_SESSION['id_utilisateur'] = $utilisateur->id_utilisateur;
                 $_SESSION['email'] = $utilisateur->email;
                 $_SESSION['role'] = $utilisateur->role;
                 $_SESSION['nom'] = $utilisateur->nom;
                 $_SESSION['prenom'] = $utilisateur->prenom;
                 $_SESSION['ville'] = $utilisateur->ville;
+                $_SESSION['crsf_token'] = bin2hex(random_bytes(32));
                 if ($_SESSION['role'] == 1) {
                     header('Location: index.php?controller=Admin&action=backoffice');
                 } else {
@@ -153,7 +163,7 @@ class UtilisateurController extends Controller
     public function disconnect()
     {
         if (isset($_SESSION['id_utilisateur'])) {
-            session_destroy();
+            session_unset();
             $_SESSION['message'] = "Vous êtes déconnecté avec succès.";
             header('Location: index.php?controller=Home&action=homeAction');
             exit();
@@ -173,6 +183,10 @@ class UtilisateurController extends Controller
 
         if ($role === 0) {
             $id_utilisateur = $_SESSION['id_utilisateur'];
+            if (empty($id_utilisateur)) {
+                header('Location: index.php?controller=Utilisateur&action=connectForm');
+                exit();
+            }
             $utilisateurModel = new UtilisateurModel();
             $utilisateur = $utilisateurModel->displayOne($id_utilisateur);
             $avisModel = new AvisModel();
@@ -199,31 +213,57 @@ class UtilisateurController extends Controller
             $id_utilisateur = $_SESSION['id_utilisateur'];
         }
 
+        $id_utilisateur = $_SESSION['id_utilisateur'];
+        $id_utilisateur_get = $_GET['id_utilisateur'] ?? $id_utilisateur;
+        if ($id_utilisateur != $id_utilisateur_get) {
+            header('Location: index.php?controller=Utilisateur&action=showProfile');
+            exit();
+        }
+
         if (!isset($id_utilisateur)) {
-            // Si l'utilisateur n'est pas connecté, rediriger vers la page de connexion
-            // $_SESSION['message'] = "Veuillez vous connecter pour modifier votre profil";
             header('Location: index.php?controller=Utilisateur&action=connectForm');
             exit;
         } else {
-            // Récupérer les informations du profil de l'utilisateur
+            $token = $_SESSION['csrf_token'] = bin2hex(random_bytes(32));
             $utilisateurModel = new UtilisateurModel();
             $utilisateur = $utilisateurModel->displayOne($id_utilisateur);
             $_SESSION['message'] = "";
-            $this->render('utilisateur/updateForm', ['utilisateur' => $utilisateur]);
+            $this->render('utilisateur/updateForm', ['utilisateur' => $utilisateur, 'token' => $token]);
         }
     }
     public function update()
     {
 
         if ($_SERVER['REQUEST_METHOD'] == 'POST') {
-            $id_utilisateur = $_POST['id_utilisateur'];
-            $nom = $_POST['name'];
-            $prenom = $_POST['surname'];
-            $numero_telephone = $_POST['tel'];
-            $ville = $_POST['city'];
-            $email = $_SESSION['email'];
+
+            if (!isset($_POST['csrf_token'], $_SESSION['csrf_token']) || !hash_equals($_SESSION['csrf_token'], $_POST['csrf_token'])){
+                header('Location: index.php?controller=Utilisateur&action=updateForm');
+                exit();
+            }
+            $id_utilisateur = filter_input(INPUT_POST, 'id_utilisateur', FILTER_VALIDATE_INT);
+            $nom = htmlspecialchars($_POST['name']);
+            $prenom = htmlspecialchars($_POST['surname']);
+            $numero_telephone = trim($_POST['tel']);
+            $ville = htmlspecialchars($_POST['city']);
+            $email = filter_var(($_POST['email']), FILTER_SANITIZE_EMAIL);
             $role = 0;
 
+            if ($role != 0) {
+                $role = 0;
+            }
+            if (empty($nom) || empty($prenom) || empty($numero_telephone) || empty($ville) || empty($email)) {
+                $_SESSION['message'] = "Veuillez remplir tous les champs.";
+                header('Location: index.php?controller=Utilisateur&action=updateForm');
+                exit();
+            }
+            $utilisateurModel = new UtilisateurModel();
+            $testMail = $utilisateurModel->displayOne($id_utilisateur);
+
+            if ($email !== $testMail->email() && $utilisateurModel->checkEmailExists($email)) {
+                $_SESSION['message'] = "L'email est déjà utilisé par un autre utilisateur.";
+                header('Location: index.php?controller=Utilisateur&action=updateForm');
+                exit();
+            }
             $utilisateur = new Utilisateur();
             $utilisateur->setId_utilisateur($id_utilisateur);
             $utilisateur->setNom($nom);
@@ -233,20 +273,16 @@ class UtilisateurController extends Controller
             $utilisateur->setEmail($email);
             $utilisateur->setRole($role);
 
-
-            $utilisateurModel = new UtilisateurModel();
-
             if ($utilisateurModel->update($utilisateur)) {
+                $_SESSION['csrf_token'] = bin2hex(random_bytes(32));
                 $_SESSION['message'] = "";
                 header('Location: index.php?controller=Utilisateur&action=showProfile');
             } else {
-                //$_SESSION['message'] = "Erreur lors de la modification";
+                $_SESSION['message'] = "Erreur lors de la modification";
                 $this->render('utilisateur/updateForm', ['utilisateur' => $utilisateur]);
             }
         }
     }
-
-    //function pour mettre à jour l'email et le mot de passe (séparée)
 
 
     public function deletePage()
@@ -257,7 +293,8 @@ class UtilisateurController extends Controller
         }
 
         if (isset($_GET['id_utilisateur']) && intval($_GET['id_utilisateur']) === $_SESSION['id_utilisateur']) {
-            $this->render('utilisateur/deletePage');
+            $token = $_SESSION['csrf_token'] = bin2hex(random_bytes(32));
+            $this->render('utilisateur/deletePage', ['token' => $token]);
         } else {
             header('Location: index.php?controller=Utilisateur&action=connectForm');
             exit();
@@ -271,9 +308,17 @@ class UtilisateurController extends Controller
             exit();
         }
 
+        if (!isset($_POST['csrf_token'], $_SESSION['csrf_token']) || !hash_equals($_SESSION['csrf_token'], $_POST['csrf_token'])) {
+            header('Location: index.php?controller=Utilisateur&action=deletePage');
+            exit();
+        }
+
         if (isset($_GET['id_utilisateur']) && intval($_GET['id_utilisateur']) === $_SESSION['id_utilisateur']) {
             $id_utilisateur = intval($_GET['id_utilisateur']);
             $utilisateurModel = new UtilisateurModel();
+
+            $demandeModel = new Demande_ReservationModel();
+            $demandeModel->deleteByIdUser($id_utilisateur);
 
             if ($utilisateurModel->delete($id_utilisateur)) {
                 $_SESSION['message'] = "Nous espérons vous revoir bientôt !";
@@ -299,22 +344,28 @@ class UtilisateurController extends Controller
         }
 
         $id_utilisateur = $_SESSION['id_utilisateur'];
+        $token = $_SESSION['csrf_token'] = bin2hex(random_bytes(32));
         $utilisateurModel = new UtilisateurModel();
         $utilisateur = $utilisateurModel->displayOne($id_utilisateur);
         if (!$utilisateur) {
             $_SESSION['message'] = "Utilisateur introuvable.";
-            header('Location: index.php?controller=Utilisateur&action=connectForm');
+            header('Location: index.php?controller=Utilisateur&action=showProfile');
             exit();
         }
-        $this->render('utilisateur/resetPasswordVerif', ['utilisateur' => $utilisateur]);
+        $this->render('utilisateur/resetPasswordVerif', ['utilisateur' => $utilisateur, 'token' => $token]);
     }
 
     public function resetPassword()
     {
         if ($_SERVER['REQUEST_METHOD'] === 'POST') {
+
+            if (!isset($_POST['csrf_token'], $_SESSION['csrf_token']) || !hash_equals($_SESSION['csrf_token'], $_POST['csrf_token'])) {
+                header('Location: index.php?controller=Utilisateur&action=resetPasswordVerif');
+                exit();
+            }
             $id_utilisateur = $_SESSION['id_utilisateur'];
-            $current_password = $_POST['current_password'];
-            $new_password = $_POST['password'];
+            $current_password = trim($_POST['current_password']);
+            $new_password = trim($_POST['password']);
 
             $utilisateurModel = new UtilisateurModel();
             $utilisateur = $utilisateurModel->displayOne($id_utilisateur);
